♻️ refactor: filter에 있던코드 home으로 이동
- storybook 제거
