+++ src/components/home/Filter/Filter.tsx
@@ -1,33 +1,14 @@
-import React, { useEffect, useState } from "react";
+import React, { useState } from "react";
 import CHANNELS from "../../../constants/channel";
-import { postAPI } from "../../../utils/apis";
 import * as S from "./style";
 
 interface FilterInterface {
-  setPost: (value: object) => void;
+  channelChange: (_id: string) => void;
 }
 
-const Filter: React.FC<FilterInterface> = ({ setPost }) => {
+const Filter: React.FC<FilterInterface> = ({ channelChange }) => {
   const channels = Object.entries(CHANNELS);
   const [channel, setChannel] = useState<string>("front");
-
-  const channelChange = (e, id: string) => {
-    setChannel(e.target.value);
-    channelSearch(id);
-  };
-
-  const channelSearch = async (id) => {
-    try {
-      const { data } = await postAPI.getChannelPostList(id);
-      setPost(data);
-    } catch (error) {
-      console.error(error);
-    }
-  };
-
-  useEffect(() => {
-    channelSearch(CHANNELS.front._id);
-  }, []);
 
   return (
     <S.FilterWrap>
@@ -40,7 +21,10 @@ const Filter: React.FC<FilterInterface> = ({ setPost }) => {
                 id={name}
                 name="channel"
                 value={name}
-                onChange={() => channelChange(event, _id)}
+                onChange={(e) => {
+                  setChannel(e.target.value);
+                  channelChange(_id);
+                }}
               />
               {channel === name ? (
                 <S.SelectLabel htmlFor={name}>{label}</S.SelectLabel>
